% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{autocompboost-package}
\alias{autocompboost}
\alias{autocompboost-package}
\title{autocompboost: AutoML with compboost}
\description{
AutoML with compboost.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coorsaa/autocompboost}
  \item Report bugs at \url{https://github.com/coorsaa/autocompboost/issues}
}

}
\author{
\strong{Maintainer}: Daniel Schalk \email{schalkdaniel242@gmail.com}

}
