% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_set.R
\name{autocompboost_default_params}
\alias{autocompboost_default_params}
\title{autocompboost_default_params}
\usage{
autocompboost_default_params()
}
\value{
\code{\link[paradox:ParamSet]{ParamSet}}
}
\description{
Default tuning search space for autocompboost.
}
