% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_pipeline.R
\name{autocompboost_preproc_pipeline}
\alias{autocompboost_preproc_pipeline}
\title{autocompboost_preproc_pipeline}
\usage{
autocompboost_preproc_pipeline(task, max_cardinality = 100)
}
\arguments{
\item{task}{(\code{\link[mlr3:Task]{Task}}) \cr
Contains the task to be solved. Currently \code{\link[mlr3:TaskClassif]{TaskClassif}} and \code{\link[mlr3:TaskRegr]{TaskRegr}} are supported.}

\item{max_cardinality}{\code{integer(1)} \cr
Maximum number of factor levels allowed. Factors with a cardinality > max_cardinality are
collapsed using \code{\link[mlr3pipelines:mlr_pipeops_collapsefactors]{PipeOpCollapseFactors}}. Default is set to 1000.}
}
\value{
\code{\link[mlr3pipelines:Graph]{Graph}}
}
\description{
AutoML pipeline of autocompboost.
}
