% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrCompboost.R
\name{mlr_learners_regr.compboost}
\alias{mlr_learners_regr.compboost}
\alias{LearnerRegrCompboost}
\title{CompBoost Regression Learner}
\description{
Componentwise boosting
Calls \code{\link[compboost:Compboost]{compboost::Compboost()}} from package \CRANpkg{compboost}.

Create a \code{LearnerRegrCompboost} object.
}
\section{Dictionary}{

This \link{Learner} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_learners} or with the associated sugar function \code{\link[=lrn]{lrn()}}:\preformatted{mlr_learners$get("regr.compboost")
lrn("regr.compboost")
}
}

\examples{
if (requireNamespace("compboost")) {
  learner = mlr3::lrn("regr.compboost")
  print(learner)

  # available parameters:
learner$param_set$ids()
}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
