% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpCVPredictionTarget.R
\name{mlr_pipeops_cv_prediction_target}
\alias{mlr_pipeops_cv_prediction_target}
\alias{PipeOpCVPredictionTarget}
\title{Wrap a Learner into a PipeOp with Cross-validated Predictions as Features}
\format{
\code{\link{R6Class}} object inheriting from \code{\link[mlr3pipelines:PipeOpTaskPreproc]{mlr3pipelines::PipeOpTaskPreproc}}/\code{\link[mlr3pipelines:PipeOp]{mlr3pipelines::PipeOp}}.
}
\description{
Wraps an \code{\link[mlr3:Learner]{mlr3::Learner}} into a \code{\link[mlr3pipelines:PipeOp]{mlr3pipelines::PipeOp}}.
}
\section{Methods}{

Methods inherited from \code{\link[mlr3pipelines:PipeOpTaskPreproc]{mlr3pipelines::PipeOpTaskPreproc}}/\code{\link[mlr3pipelines:PipeOp]{mlr3pipelines::PipeOp}}.
}

\seealso{
https://mlr3book.mlr-org.com/list-pipeops.html
}
