% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpExtractIntercations.R
\name{mlr_pipeops_extract_interactions}
\alias{mlr_pipeops_extract_interactions}
\alias{PipeOpExtractInteractions}
\title{PipeOpExtractInteractions}
\format{
Abstract \code{\link{R6Class}} inheriting from \code{\link[mlr3pipelines:PipeOp]{PipeOp}}.
}
\description{
PipeOperator to extract interactions gained by an \code{rpart} model.
}
\section{Construction}{
\preformatted{PipeOpExtractInteractions$new(id = "extract_interactions", param_vals = list())
}
\itemize{
\item \code{id} :: \code{character(1)}\cr
Identifier of resulting object, default \code{"encode"}.
\item \code{param_vals} :: named \code{list}\cr
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default \code{list()}.
}
}

\section{Input and Output Channels}{

Input and output channels are inherited from \code{\link[mlr3pipelines:PipeOp]{PipeOp}}.

Based on the input \code{\link[mlr3:Task]{Task}}, a tree model is trained.
The output contains interactions as \code{list()}.
}

\section{State}{

The \verb{$state} is a named \code{list} with the \verb{$state} elements inherited from \code{\link[mlr3pipelines:PipeOp]{PipeOp}}.
}

\section{Parameters}{

The parameters are the parameters inherited from \code{\link[mlr3pipelines:PipeOp]{PipeOp}}, as well as:
\itemize{
\item \code{degree}  :: \code{integer(1)} \cr
Degree of interactions. Initialized to \code{2L}.
}
}

\section{Methods}{

Only methods inherited from \code{\link[mlr3pipelines:PipeOp]{PipeOp}}.
}

\concept{PipeOps}
