% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{autocompboost-package}
\alias{autocompboost}
\alias{autocompboost-package}
\title{autocompboost: AutoML with compboost}
\description{
AutoML with compboost.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coorsaa/autocompboost}
  \item Report bugs at \url{https://github.com/coorsaa/autocompboost/issues}
}

}
\author{
\strong{Maintainer}: Stefan Coors \email{coors.stefan@gmail.de} (\href{https://orcid.org/0000-0002-7465-2146}{ORCID})

Authors:
\itemize{
  \item Daniel Schalk \email{schalkdaniel242@gmail.com}
}

}
