% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distilledRF.R
\name{distilledRF}
\alias{distilledRF}
\title{RF to CART distiller}
\usage{
distilledRF(task, max_time)
}
\description{
Calls \code{\link[rpart:rpart]{rpart::rpart()}} to train a decision tree on cross-validated predictions of a
random forest using \code{\link[ranger:ranger]{ranger::ranger()}} from package \CRANpkg{ranger}.
Cross-validated predictions are obtained by \link[mlr3pipelines:mlr_pipeops_learner_cv]{mlr3pipelines::PipeOpLearnerCV}.
}
